(1) Add new CT project, CTjsonlib; (2) Move PlayerComponentState.java and PlayerWorldState.java from CTtraveler project to CTjsonlib project; (3) Update NOTICE file to include statement about use of the Google Gson library
