+++ JavaCode/CTjsonlib/build.gradle
@@ -1,0 +1,12 @@
+
+dependencies {
+	compile 'com.google.code.gson:gson:2.8.5'
+}
+
+// JAR settings
+jar {
+    // Include license in the JAR's META-INF directory
+    into("META-INF") {
+    	from "$rootDir/../LICENSE", "$rootDir/../NOTICE" 
+    }
+}
