+++ JavaCode/CTtraveler/build.gradle
@@ -37,14 +37,19 @@ repositories {
     flatDir {
     	// Location of CTlib.jar
     	dirs "$rootProject.projectDir/CTlib/build/libs"
+        // Location of CTjsonlib.jar
+        dirs "$rootProject.projectDir/CTjsonlib/build/libs"
     }
 }
 
 dependencies {
 	// Build CTlib project first, because we use CTlib.jar
 	compile project(':CTlib')
+    // Build CTjsonlib project first, because we use CTjsonlib.jar
+    compile project(':CTjsonlib')
 	// Specify that we use CTlib.jar
     compile name: 'CTlib'
+    // Specify that we use CTjsonlib.jar
+    compile name: 'CTjsonlib'
     compile group: 'commons-cli', name: 'commons-cli', version: '1.3.1'
-    compile 'com.google.code.gson:gson:2.8.5'
 }
